+++ CHANGELOG.md
@@ -25,6 +25,7 @@ https://github.com/coleifer/peewee/releases
 * Adds new `ISODateTimeField` for Sqlite that encodes datetimes in ISO format
   (more friendly when db is shared with other tools), and also properly reads
   back UTC offset info.
+* Remove `playhouse.sqlite_ext.ClosureTable` implementation.
 * Add a `Model.dirty_field_names` attribute that is safe for membership
   testing, since testing `x in dirty_fields` returns True if one or more field
   exists due to operator overloads returning a truthy Expression object.